[Chore/#172] 코드 정리
